requirements-dev.txt
Pyrogram==2.0.97
TgCrypto==1.2.5
uvloop
loguru
python-dotenv
black







